+++ Plugin/src/Patches.cs
@@ -14,6 +14,7 @@ using CodeRebirth.WeatherStuff;
 using UnityEngine.AI;
 using UnityEngine.SceneManagement;
 using Random = UnityEngine.Random;
+using System.Linq;
 
 namespace CodeRebirth.src;
 [HarmonyPatch(typeof(StartOfRound))]
@@ -32,7 +33,11 @@ internal static class StartOfRoundPatcher {
     public static void StartOfRound_Start(ref StartOfRound __instance)
     {
         __instance.NetworkObject.OnSpawn(CreateNetworkManager);
-        LethalLib.Modules.Weathers.RemoveWeather("Meteor Shower", LethalLib.Modules.Levels.LevelTypes.None, [Plugin.ModConfig.ConfigMeteorShowerMoonsBlacklist.Value]);
+
+        string[] levelOverrides = Plugin.ModConfig.ConfigMeteorShowerMoonsBlacklist.Value.Split(',')
+                                    .Select(name => name.Trim())
+                                    .ToArray();
+        LethalLib.Modules.Weathers.RemoveWeather("Meteor Shower", levelOverrides: levelOverrides);
     }
 
     [HarmonyPatch(nameof(StartOfRound.OnDisable))]